+++ app/controllers/base_public_controller.rb
@@ -1,5 +1,5 @@
 class BasePublicController < ApplicationController
-  CLAIM_TIMEOUT_LENGTH_IN_MINUTES = 30
+  include ClaimSessionTimeout
 
   helper_method :current_policy_routing_name, :claim_timeout_in_minutes
   before_action :end_expired_claim_sessions
@@ -8,10 +8,6 @@ class BasePublicController < ApplicationController
 
   def current_policy_routing_name
     params[:policy]
-  end
-
-  def claim_timeout_in_minutes
-    self.class::CLAIM_TIMEOUT_LENGTH_IN_MINUTES
   end
 
   def end_expired_claim_sessions
@@ -25,15 +21,5 @@ class BasePublicController < ApplicationController
         end
       end
     end
-  end
-
-  def claim_session_timed_out?
-    session.key?(:claim_id) && session[:last_seen_at] < claim_timeout_in_minutes.minutes.ago
-  end
-
-  def clear_claim_session
-    session.delete(:claim_id)
-    session.delete(:verify_request_id)
-    @current_claim = nil
   end
 end
